Many changes related to getting the scotland mode version running. Some fixes being reapplied after they mysteriously vanished during the great debacle that preceded the creation of the development branch. Not that Im bitter
