+++ minos/modules/housing.py
@@ -93,7 +93,7 @@ class Housing(Base):
         # TODO: Find out why this was adding 1 to the prediction?
         #housing_prob_df["housing_quality"] = self.random.choice(housing_prob_df.index, list(housing_prob_df.columns), housing_prob_df)+1
         housing_prob_df["housing_quality"] = self.random.choice(housing_prob_df.index, list(housing_prob_df.columns),
-                                                                housing_prob_df)
+                                                                housing_prob_df) + 1
         housing_prob_df.index = housing_prob_df.index.astype(int)
 
         self.population_view.update(housing_prob_df["housing_quality"])
