+++ minos/modules/loneliness.py
@@ -87,13 +87,13 @@ class Loneliness(Base):
 
         loneliness_prob_df = self.calculate_loneliness(pop)
 
-        #loneliness_prob_df["loneliness"] = self.random.choice(loneliness_prob_df.index,
-        #                                                      list(loneliness_prob_df.columns),
-        #                                                      loneliness_prob_df) + 1
-
         loneliness_prob_df["loneliness"] = self.random.choice(loneliness_prob_df.index,
                                                               list(loneliness_prob_df.columns),
-                                                              loneliness_prob_df)
+                                                              loneliness_prob_df) + 1
+
+        #loneliness_prob_df["loneliness"] = self.random.choice(loneliness_prob_df.index,
+        #                                                      list(loneliness_prob_df.columns),
+        #                                                      loneliness_prob_df)
 
         loneliness_prob_df.index = loneliness_prob_df.index.astype(int)
